Refactor folder and file structure to limit sidenav to the dashboard, not login page.
