+++ src/app/(dashboard)/weather/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { WeeklyComponent } from "@/components/WeeklyComponent";
 import SearchBar from "@/components/SearchBar";
-import { DisplayedCityWeatherProvider } from "../contexts/DisplayedCityWeatherContext";
+import { DisplayedCityWeatherProvider } from "../../../contexts/DisplayedCityWeatherContext";
 
 export default function Home() {
   return (
